+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Checker script to read in APDDs and check if they are compatible with the library
+### Changed
+- CPX-AP now interprets apdds with utf-8 instead of ascii encoding
 
 ## v0.5.1 - 21.06.24
 ### Fixed